requirements.txt
# Core dependencies
flask>=3.0.0
requests>=2.31.0
pyyaml>=6.0.1
asyncio>=3.4.3

# Security and HTTP
urllib3>=2.0.7
certifi>=2023.7.22

# Data processing
numpy>=1.24.3
pandas>=2.0.3

# Logging and monitoring
colorama>=0.4.6
tqdm>=4.66.1

# Development and testing
pytest>=7.4.2
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1

# Documentation (optional)
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0

# Code quality
flake8>=6.0.0
black>=23.7.0
mypy>=1.5.1

# Web interface extras
jinja2>=3.1.2
werkzeug>=3.0.1

# Async support
aiohttp>=3.8.5
aiofiles>=23.2.1

# JSON handling
jsonschema>=4.19.0

# Time and date
python-dateutil>=2.8.2

# Regular expressions
regex>=2023.8.8
